app.js fix
